+++ src/DependencyInjection/MeritooCommonExtension.php
@@ -26,9 +26,6 @@ class MeritooCommonExtension extends BaseExtension
      */
     protected function getBundleDirectoryPath(): string
     {
-        return Miscellaneous::concatenatePaths([
-            __DIR__,
-            '..',
-        ]);
+        return Miscellaneous::concatenatePaths(__DIR__, '..');
     }
 }